Fix MCP server path
